Update get_headers.py
Trying to deal with the COMMENTS header and the fact that it (1) possibly has blanks and (2) that its more than one header in a row with the same name. Goal is to first print out all of the headers from a FITS file. Then go through a process where we assign each header we find to a variable. Then we should print all of the contents of each variable out to see what was created in terms of variable names and their assigned values.